ACP-2626 Refactored.
