+++ src/Spryker/Zed/AppMerchant/Business/AppMerchantBusinessFactory.php
@@ -13,7 +13,7 @@ use Spryker\Zed\AppMerchant\Business\MerchantAppOnboarding\MerchantAppOnboarding
 use Spryker\Zed\AppMerchant\Business\MerchantAppOnboarding\Writer\MerchantAppOnboardingCreator;
 use Spryker\Zed\AppMerchant\Business\MerchantAppOnboarding\Writer\MerchantAppOnboardingCreatorInterface;
 use Spryker\Zed\AppMerchant\Business\Message\MessageSender;
-use Spryker\Zed\AppMerchant\Business\PaymentsTransmissions\PaymentTransmissionsRequestExtender;
+use Spryker\Zed\AppMerchant\Business\PaymentTransmissions\PaymentTransmissionsRequestExtender;
 use Spryker\Zed\AppMerchant\Business\Webhook\WebhookHandler;
 use Spryker\Zed\AppMerchant\Business\Webhook\WebhookRequestExtender;
 use Spryker\Zed\AppMerchant\Dependency\Facade\AppMerchantToAppKernelFacadeInterface;
